add generic extensions for IDependencyResolver
and extract DI container tests into their own files
